requirements.txt
python-dotenv==1.0.0
smtplib
email
